backend/ml_integration/image_model.py
import numpy as np
import cv2
import logging
import torch
from PIL import Image
import os

logger = logging.getLogger(__name__)

# Placeholder for a real ML model - would be replaced with an actual trained model
class ImageAnalysisModel:
    def __init__(self):
        """Initialize the image analysis model"""
        self.model_loaded = False
        try:
            # This would be replaced with actual model loading code
            # Example: self.model = torch.load('model_path.pth')
            self.model_loaded = True
            logger.info("Image analysis model loaded successfully")
        except Exception as e:
            logger.error(f"Error loading image analysis model: {e}")
            
    def preprocess_image(self, image_path):
        """Preprocess the image for the model"""
        try:
            # Read the image
            img = cv2.imread(image_path)
            if img is None:
                logger.error(f"Could not read image at path: {image_path}")
                return None
                
            # Resize to model input size
            img_resized = cv2.resize(img, (224, 224))
            
            # Convert to RGB if needed
            img_rgb = cv2.cvtColor(img_resized, cv2.COLOR_BGR2RGB)
            
            # Normalize
            img_normalized = img_rgb / 255.0
            
            # Convert to tensor (would be required for a real model)
            # Example: img_tensor = torch.from_numpy(img_normalized).permute(2, 0, 1).float()
            
            return img_normalized
        except Exception as e:
            logger.error(f"Error preprocessing image: {e}")
            return None
            
    def predict(self, preprocessed_image):
        """Run inference on the preprocessed image"""
        # This is a placeholder for real model prediction
        # In a real implementation, this would use the loaded ML model
        
        # For demo purposes, return a random prediction
        authenticity_score = np.random.uniform(0, 1)
        anomaly_map = np.random.rand(224, 224)  # Random heatmap
        
        # Determine areas with highest anomaly scores
        anomaly_threshold = 0.7
        anomaly_coords = np.where(anomaly_map > anomaly_threshold)
        
        anomalies = []
        for y, x in zip(anomaly_coords[0], anomaly_coords[1]):
            # Convert to original image coordinates
            anomalies.append({
                'x': int(x),
                'y': int(y),
                'severity': float(anomaly_map[y, x]),
                'type': 'surface_defect' if anomaly_map[y, x] > 0.8 else 'marking_issue'
            })
        
        return {
            'authenticity_score': float(authenticity_score),
            'anomaly_map': anomaly_map.tolist() if len(anomalies) > 0 else None,
            'anomalies': anomalies
        }

# Create a singleton instance
_model_instance = None

def get_model():
    """Get or create the model instance"""
    global _model_instance
    if _model_instance is None:
        _model_instance = ImageAnalysisModel()
    return _model_instance

def analyze_image(image_path):
    """
    Analyze an image to detect counterfeit indicators
    
    Args:
        image_path: Path to the component image
        
    Returns:
        Dict with analysis results including:
        - confidence: Confidence score (0-1)
        - classification: PASS/SUSPECT/FAIL
        - anomaly_map: Heatmap of anomalies
        - anomalies: List of anomaly coordinates and types
    """
    try:
        model = get_model()
        
        # Preprocess the image
        preprocessed_image = model.preprocess_image(image_path)
        if preprocessed_image is None:
            return {
                'confidence': 0,
                'classification': 'FAIL',
                'error': 'Failed to process image'
            }
        
        # Make prediction
        prediction = model.predict(preprocessed_image)
        
        # Determine classification based on authenticity score
        authenticity_score = prediction['authenticity_score']
        
        if authenticity_score > 0.8:
            classification = 'PASS'
        elif authenticity_score > 0.5:
            classification = 'SUSPECT'
        else:
            classification = 'FAIL'
        
        return {
            'confidence': authenticity_score,
            'classification': classification,
            'anomaly_map': prediction.get('anomaly_map'),
            'anomalies': prediction.get('anomalies', [])
        }
    
    except Exception as e:
        logger.error(f"Error in image analysis: {e}")
        return {
            'confidence': 0,
            'classification': 'FAIL',
            'error': str(e)
        }